Fixed Auth.isAuthorized

+++ server/api/tangos/tangos.routes.js
@@ -22,12 +22,17 @@ router.get('/', Auth.hasRole('admin'), function(req, res) {
   ;
 });
 
-router.get('/:id', Auth.isAuthorized, function(req, res) {
+router.get('/:id', Auth.isLoggedIn, function(req, res) {
   Tango.findById(req.params.id).exec()
     .then(function(tango) {
       if (!tango) {
         return res.status(404).end();
       }
+
+      if (Auth.isAuthorized(tango.author, req.user._id, req, res)) {
+        return;
+      }
+
       return res.status(200).json(tango);
     }, forwardError(res))
   ;
@@ -38,6 +43,8 @@ router.post('/', Auth.isLoggedIn, function(req, res) {
   // 2. Find current user
   // 3. Add tango to current user
   // 4. Save current user
+
+  req.body.author = req.user._id;
   Tango.create(req.body)
     .then(function(createdTango) {
       User.findById(req.user._id).exec()
@@ -58,12 +65,17 @@ router.post('/', Auth.isLoggedIn, function(req, res) {
   ;
 });
 
-router.put('/:id', Auth.isAuthorized, function(req, res) {
+router.put('/:id', Auth.isLoggedIn, function(req, res) {
   Tango.findByIdAndUpdate(req.params.id, req.body, { new: true, runValidators: true }).exec()
     .then(function(tango) {
       if (!tango) {
         return res.status(404).end();
       }
+
+      if (Auth.isAuthorized(tango.author, req.user._id, req, res)) {
+        return;
+      }
+
       return res.status(201).json(tango);
     })
     .then(null, function(err) { // invalid tango
@@ -72,26 +84,35 @@ router.put('/:id', Auth.isAuthorized, function(req, res) {
   ;
 });
 
-router.delete('/:id', Auth.isAuthorized, function(req, res) {
+router.delete('/:id', Auth.isLoggedIn, function(req, res) {
+  // 0. Find tango to check if authorized
   // 1. Remove tango
   // 2. Find current user
   // 3. Remove tango from current user
   // 4. Save current user
-  Tango.findByIdAndRemove(req.params.id).exec()
-    .then(function(removedTango) {
-      if (!removedTango) {
+  Tango.findById(req.params.id).exec()
+    .then(function(foundTango) {
+      if (!foundTango) {
         return res.status(404).end();
       }
-      User.findById(req.user._id).exec()
-        .then(function(currentUser) {
-          var index = currentUser.tangos.indexOf(removedTango);
-          currentUser.tangos.splice(index, 1);
-          currentUser.save(function(err) {
-            if (err) {
-              return res.status(500).send({ error: err });
-            }
-            res.status(204).end();
-          });
+      if (Auth.isAuthorized(foundTango.author, req.user._id, req, res)) {
+        return;
+      }
+
+      Tango.findByIdAndRemove(req.params.id).exec()
+        .then(function(removedTango) {
+          User.findById(req.user._id).exec()
+            .then(function(currentUser) {
+              var index = currentUser.tangos.indexOf(removedTango);
+              currentUser.tangos.splice(index, 1);
+              currentUser.save(function(err) {
+                if (err) {
+                  return res.status(500).send({ error: err });
+                }
+                res.status(204).end();
+              });
+            }, forwardError(res))
+          ;
         }, forwardError(res))
       ;
     }, forwardError(res))
